# docs/frontend/openapi-spec.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "InsightLogger.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/providers/ai": {
      "get": {
        "tags": ["Providers"],
        "summary": "Return configured AI providers and normalized capability metadata.",
        "operationId": "GetAiProviders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAiProvidersResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiProviderCapabilitiesContract": {
        "type": "object",
        "properties": {
          "supportsStreaming": {
            "type": "boolean"
          },
          "supportsToolCalling": {
            "type": "boolean"
          },
          "supportsJsonMode": {
            "type": "boolean"
          },
          "supportsOpenAiCompatibility": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "AiProviderItemContract": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "defaultModel": {
            "type": "string",
            "nullable": true
          },
          "capabilities": {
            "$ref": "#/components/schemas/AiProviderCapabilitiesContract"
          }
        },
        "additionalProperties": false
      },
...
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            },
            "nullable": true
          },
          "correlationId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "additionalProperties": false
      },
...
      "GetAiProvidersResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiProviderItemContract"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationErrorDetail": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}